// src/app/models/usuario.ts
export class Usuario {
  idUsuario: number = 0;
  nombreUsuario: string = '';
  apellidoUsuario: string = '';
  correoUsuario: string = '';
  telefonoUsuario: string = '';
  direccionUsuario: string = '';
  estadoUsuario: boolean = false;
  usuarioUsuario: string = '';
  contraseniaUsuario: string = '';
  favoritoUsuario: boolean = false;
  edadUsuario: number = 0;
}